#1824 break exising ModuleInlineTrans test out into new file
